+++ src/Views/UserControlTallyCounter.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace MagonoteToolkitForEmbedded.Views
+{
+    /// <summary>
+    /// Interaction logic for UserControlTallyCounter
+    /// </summary>
+    public partial class UserControlTallyCounter : UserControl
+    {
+        public UserControlTallyCounter()
+        {
+            InitializeComponent();
+        }
+    }
+}